Restructure command code
