+++ canvas/cli/__init__.py
@@ -4,4 +4,3 @@ from __future__ import annotations
 from .parse import *
 from .factory import *
 from .base import *
-from .init import *
