+++ canvas/cli/commands/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import annotations
+
+# flake8: noqa
+from .init import *
